# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/api/move": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Получить список ходов с возможностью фильтрации по статусу и датам",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "moves"
                ],
                "summary": "Получить все заявки на ходы с параметрами",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Статус хода",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Наличие статуса",
                        "name": "is_status",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "format": "date",
                        "description": "Дата от",
                        "name": "from_date",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "format": "date",
                        "description": "Дата до",
                        "name": "to_date",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/schemas.GetAllMovesWithParamsResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/schemas.ResponseMessage"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/schemas.ResponseMessage"
                        }
                    }
                }
            }
        },
        "/api/move/finish/{ID}": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Завершить ход по его ID, обновив его статус",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "moves"
                ],
                "summary": "Завершить ход",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID хода",
                        "name": "ID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Статус завершенного хода",
                        "name": "status",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Move was Finished",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/schemas.ResponseMessage"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/schemas.ResponseMessage"
                        }
                    }
                }
            }
        },
        "/api/move/form/{ID}": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Сформировать ход по его ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "moves"
                ],
                "summary": "Сформировать ход",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID хода",
                        "name": "ID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Move was Formed",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/schemas.ResponseMessage"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/schemas.ResponseMessage"
                        }
                    }
                }
            }
        },
        "/api/move/{ID}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Получить детализированную информацию о ходе, включая карты, участвующие в ходе, и их статус",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "moves"
                ],
                "summary": "Получить заявку на ход по ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID хода",
                        "name": "ID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/schemas.GetMoveResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/schemas.ResponseMessage"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/schemas.ResponseMessage"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/schemas.ResponseMessage"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Обновить информацию о ходе, включая поля игрока и этапа",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "moves"
                ],
                "summary": "Обновить поля хода",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID хода",
                        "name": "ID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Имя игрока",
                        "name": "player",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Этап хода",
                        "name": "stage",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Fields was updated",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/schemas.ResponseMessage"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/schemas.ResponseMessage"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Удалить заявку на ход по ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "moves"
                ],
                "summary": "Удалить ход",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID хода",
                        "name": "ID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Move was deleted",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/schemas.ResponseMessage"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/schemas.ResponseMessage"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "ds.Cards": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "image_url": {
                    "type": "string"
                },
                "long_description": {
                    "type": "string"
                },
                "multiplier": {
                    "type": "string"
                },
                "status": {
                    "type": "boolean"
                },
                "title_en": {
                    "type": "string"
                },
                "title_ru": {
                    "type": "string"
                }
            }
        },
        "ds.Moves": {
            "type": "object",
            "properties": {
                "creator": {
                    "$ref": "#/definitions/ds.Users"
                },
                "creator_id": {
                    "type": "integer"
                },
                "cube": {
                    "type": "integer"
                },
                "date_create": {
                    "type": "string"
                },
                "date_finish": {
                    "type": "string"
                },
                "date_update": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "moderator": {
                    "$ref": "#/definitions/ds.Users"
                },
                "moderator_id": {
                    "type": "integer"
                },
                "player": {
                    "type": "string"
                },
                "stage": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                }
            }
        },
        "ds.Users": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "is_moderator": {
                    "type": "boolean"
                },
                "login": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
...
        "schemas.GetAllMovesWithParamsResponse": {
            "type": "object",
            "properties": {
                "moves": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ds.Moves"
                    }
                }
            }
        },
...
        "schemas.GetMoveResponse": {
            "type": "object",
            "properties": {
                "move_cards": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/schemas.InfoForMove"
                    }
                },
                "moves": {
                    "type": "object",
                    "additionalProperties": true
                }
            }
        },
        "schemas.InfoForMove": {
            "type": "object",
            "properties": {
                "card": {
                    "$ref": "#/definitions/ds.Cards"
                },
                "food": {
                    "type": "integer"
                }
            }
        },
...
        "schemas.ResponseMessage": {
            "type": "object"
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}